pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain-groq>=0.3.7",
    "langchain[all]>=0.3.27",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
